+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 ### Changed
+- minimum IntelliJ platform version supported is now 2023.3.1
 
 ### Deprecated